+++ src/pbrenamer/ui/settings_dialog.py
@@ -33,6 +33,8 @@ class SettingsDialog(QDialog):
             self._ui.cmbLogLevel.setCurrentIndex(idx)
 
         self._ui.chkRestoreLastDir.setChecked(settings.get_restore_last_dir())
+        self._ui.chkRestoreToolbarState.setChecked(settings.get_restore_toolbar_state())
+        self._ui.spnPreviewDelay.setValue(settings.get_preview_delay())
 
         self._ui.buttonBox.accepted.connect(self._save_and_accept)
 
@@ -42,4 +44,6 @@ class SettingsDialog(QDialog):
         settings.set_log_level(level)
         settings.apply_log_level(level)
         settings.set_restore_last_dir(self._ui.chkRestoreLastDir.isChecked())
+        settings.set_restore_toolbar_state(self._ui.chkRestoreToolbarState.isChecked())
+        settings.set_preview_delay(self._ui.spnPreviewDelay.value())
         self.accept()